api/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.43.2",
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
    "pandas>=2.2.3",
    "sqlalchemy>=2.0.39",
    "python-pptx>=1.0.2",
]
